# docs/phase1/api/openapi.yaml
openapi: 3.0.3
info:
  title: NovelShelf API
  description: 個人利用の広告なしWeb小説リーダー「NovelShelf」のバックエンドAPI。
  version: 0.1.0-phase1
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /auth/signup:
    post:
      tags: [auth]
      summary: メールアドレス+パスワードでアカウント作成
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password, minLength: 8 }
                displayName: { type: string }
      responses:
        "201":
          description: 作成成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "409": { description: メールアドレス重複 }
  /auth/login:
    post:
      tags: [auth]
      summary: ログイン
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "401": { description: 認証失敗 }
  /auth/refresh:
    post:
      tags: [auth]
      summary: アクセストークン再発行
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
...
  /novels/resolve:
    post:
      tags: [novels]
      summary: 作品URLから作品を登録（対応サイトなら実メタデータ、未対応サイトならリンクのみのプレースホルダーを作成）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: { type: string, format: uri }
      responses:
        "200":
          description: |
            成功。対応サイト（実装時点ではなろうのみ）ならタイトル・作者・あらすじ等の実メタデータを取得して返す。
            未対応サイト（カクヨム/ハーメルン/pixiv小説）の場合は本文取得を行わず、
            URLをtitle/sourceUrlとしたプレースホルダー作品を作成して返す（本棚登録のみ可能、話一覧・本文取得は空/451になる）。
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Novel" }
        "422":
          description: URLがどの対応サイトのドメインとも一致せず、作品として認識できない
  /novels/{novelId}:
    get:
      tags: [novels]
      summary: 作品詳細取得
      parameters:
        - $ref: "#/components/parameters/NovelId"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NovelDetail" }
  /novels/{novelId}/chapters:
    get:
      tags: [novels]
      summary: 話一覧取得
      parameters:
        - $ref: "#/components/parameters/NovelId"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Chapter" }
  /novels/{novelId}/download:
    post:
      tags: [novels]
      summary: 全話本文を一括取得する（作品単位のオフライン保存用）
      description: |
        サイトへのリクエストは1話ごとに規約準拠のレート制限（なろうの場合1秒間隔）を維持したまま順次取得するため、
        話数が多い作品ほどレスポンスに時間がかかる（例: 300話で約5分）。本文はサーバーに保存されず、都度サイトから取得して返す。
      parameters:
        - $ref: "#/components/parameters/NovelId"
      responses:
        "200":
          description: 成功（全話分の本文）
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ChapterWithContent" }
        "451":
          description: 規約上、本文の自動取得ができないサイト
  /novels/search:
    get:
      tags: [novels]
      summary: タイトル/作者/ジャンル/サイト/タグで検索
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: site
          in: query
          schema: { type: string }
        - name: genre
          in: query
          schema: { type: string }
        - name: tag
          in: query
          schema: { type: string }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Novel" }
...
  /reading/positions/{novelId}:
    get:
      tags: [reading]
      summary: 作品の現在の読書位置を取得
      parameters:
        - $ref: "#/components/parameters/NovelId"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReadingPosition" }
    put:
      tags: [reading]
      summary: 読書位置を更新（クロスデバイス同期）
      parameters:
        - $ref: "#/components/parameters/NovelId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chapterId, scrollPosition]
              properties:
                chapterId: { type: string, format: uuid }
                scrollPosition: { type: number, format: float }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReadingPosition" }
  /reading/history:
    post:
      tags: [reading]
      summary: 読書履歴を記録（話を読み終えた/一定時間閲覧した時にクライアントから送信）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chapterId, readAt, durationSeconds]
              properties:
                chapterId: { type: string, format: uuid }
                readAt: { type: string, format: date-time }
                durationSeconds: { type: integer }
      responses:
        "201": { description: 成功 }
    get:
      tags: [reading]
      summary: 最近読んだ作品/話の履歴取得
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ReadingHistoryItem" }

  # ---------- Bookmarks ----------
  /bookmarks:
    get:
      tags: [bookmarks]
      summary: しおり一覧取得
      parameters:
        - name: novelId
          in: query
          schema: { type: string, format: uuid }
        - name: tag
          in: query
          schema: { type: string }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Bookmark" }
    post:
      tags: [bookmarks]
      summary: しおり作成
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookmarkInput" }
      responses:
        "201":
          description: 成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Bookmark" }
  /bookmarks/{bookmarkId}:
    patch:
      tags: [bookmarks]
      summary: しおり更新
      parameters:
        - name: bookmarkId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookmarkInput" }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Bookmark" }
    delete:
      tags: [bookmarks]
      summary: しおり削除
      parameters:
        - name: bookmarkId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: 成功 }

  # ---------- Tags ----------
...
  /offline/preferences:
    get:
      tags: [offline]
      summary: オフライン保存意図（自動/明示）の一覧をクロスデバイス取得
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OfflineSavePreference" }
    post:
      tags: [offline]
      summary: 話をオフライン保存対象として登録
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chapterId]
              properties:
                chapterId: { type: string, format: uuid }
      responses:
        "201": { description: 成功 }
  /offline/preferences/{chapterId}:
    delete:
      tags: [offline]
      summary: オフライン保存対象から除外
      parameters:
        - $ref: "#/components/parameters/ChapterId"
      responses:
        "204": { description: 成功 }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    NovelId:
      name: novelId
      in: path
      required: true
      schema: { type: string, format: uuid }
    ChapterId:
      name: chapterId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
    AuthTokens:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer }
...
    Novel:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        author: { type: string }
        site: { type: string, enum: [NAROU, KAKUYOMU, HAMELN, PIXIV] }
        genre: { type: string }
        coverUrl: { type: string, format: uri }
        sourceUrl: { type: string, format: uri }
        status: { type: string, enum: [ONGOING, COMPLETED] }
        latestKnownChapterNo: { type: integer }
        hasUpdate: { type: boolean }
    NovelDetail:
      allOf:
        - $ref: "#/components/schemas/Novel"
        - type: object
          properties:
            synopsis: { type: string }
            seriesId: { type: string, format: uuid, nullable: true }
            totalChapters: { type: integer }
    Chapter:
      type: object
      properties:
        id: { type: string, format: uuid }
        novelId: { type: string, format: uuid }
        chapterNo: { type: integer }
        title: { type: string }
        publishedAt: { type: string, format: date-time }
...
    ChapterWithContent:
      type: object
      properties:
        chapterId: { type: string, format: uuid }
        chapterNo: { type: integer }
        title: { type: string }
        bodyHtml: { type: string }
        sourceUrl: { type: string, format: uri }
...
    ReadingPosition:
      type: object
      properties:
        novelId: { type: string, format: uuid }
        chapterId: { type: string, format: uuid }
        scrollPosition: { type: number, format: float }
        lastReadAt: { type: string, format: date-time }
    ReadingHistoryItem:
      type: object
      properties:
        novelId: { type: string, format: uuid }
        chapterId: { type: string, format: uuid }
        title: { type: string }
        readAt: { type: string, format: date-time }
        durationSeconds: { type: integer }
    BookmarkInput:
      type: object
      required: [chapterId, name]
      properties:
        chapterId: { type: string, format: uuid }
        name: { type: string }
        memo: { type: string }
        scrollPosition: { type: number, format: float }
        tagIds:
          type: array
          items: { type: string, format: uuid }
    Bookmark:
      allOf:
        - $ref: "#/components/schemas/BookmarkInput"
        - type: object
          properties:
            id: { type: string, format: uuid }
            novelId: { type: string, format: uuid, description: "対象話が属する作品ID（読書画面への遷移用）" }
            chapterNo: { type: integer }
            chapterTitle: { type: string }
            createdAt: { type: string, format: date-time }
...
    OfflineSavePreference:
      type: object
      properties:
        chapterId: { type: string, format: uuid }
        autoCached: { type: boolean }
        requestedAt: { type: string, format: date-time }